Added an ansible config file
I did this so that we can always change the value of ansible_managed.
When it included a timestamp templates were reporting changed when they
were not actually changed.

This also stores the vault_password_file setting which will need to be
modified most likely when not running ansible from you own machine.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>