Move packages list into pkg manager specific var file.
We'll need to at least have lists of packages to install for yum vs apt
systems.  Maybe even eventually distro / version specific lists.

I also added in a check to make sure the packages list is defined
because it defaults to [] in defaults/main.yml and the template module
that uses this list will just silently be skipped if packages isn't
defined.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>