move Red Hat entitlements into "common" role
Prior to this commit, entitling a system with Red Hat was a part of the
testnode role.

Move the Red Hat entitlement tasks into a new "common" role so that it
can apply to other systems besides testnodes.

This will allow us to register other systems to consume updates from Red
Hat's CDN.

Also, switch the "when" clause to evalute "ansible_distribution" instead
of "ansible_os_family". Prior to this change, Ansible would include
setup-redhat.yml on RHEL, Fedora and CentOS. The ansible_os_family fact
was overly-broad there, because we do not want to run
subscription-manager on Fedora and CentOS.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>