downstream-setup: adds a new downstream_setup.yml playbook.
This playbook will be used to setup testnodes for downstream
testing by downloading arbitrary yum repos. The repos to install should
be passed in with the --extra-vars option to ansible-playbook.

The yum_repos var must be a list of hashes defining the url to download
the repo from and the name to save it as in /etc/yum.repos.d

For example:

  --extra-vars '{"yum_repos":[{"url": "<url>", "name": "name"}]}'

Signed-off-by: Andrew Schoen <aschoen@redhat.com>