puddle: use distill configs from rcm-metadata.git
Originally I thought that we could store each of distill's individual
configuration files in ceph-cm-ansible.git, but this is becoming
untenable because there are too many templated values, and I'm trying to
sync these files closely with the copies from Red Hat's rel-eng team.

This commit uses the git module to clone a rcm-metadata.git repository
to /etc/distill and alters the "ceph-distill" wrapper script to use that
location instead.

Since the individual templated files are no longer needed (the full
versions are in rcm-metadata.git), delete them from the tree here.