cobbler: keep cobbler from messing with sources.list
We want this because on vivid we want to use the distro provided sources
and when we use the cobbler snippet it wipes out those values in
sources.list. On the other versions of ubuntu we use the testnodes
playbook to provide our own sources and repos because we are currently
building packages for those, but don't want that burden for vivid.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>