testnode: enable nfs-server on rhel 7.x
In the port from ceph-qa-chef I decided to enable nfs-server by default
on all distros even though chef only enabled it for rhel 7.x. We've
since discovered that enabling it on vivid causes errors, so we're not
enabling it for all distros anymore. This makes sure we enable it still
on rhel 7.x.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>