Use a targets var to optionally define hosts for each playbook
This will allow us to pass something like --extra-vars="targets=all"
when running the testnodes playbook to allow it to use any node in the
inventory and not only those in the testnodes group.

I see this as being useful for teuthology so that when the Ansible task
creates an inventory dynamically it doesn't have to worry about trying
to match up the nodes in the inventory with what's defined in the playbook.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>