puddle: add ceph-1.2 to async configs
Add the the CEPH-1.2 release to the CEPH-ASYNC config files.

The purpose of this change is to make it easier for QE to consume
the builds in Puddle repos that correspond to ASYNC advisories.

Prior to this change, QE had to make sure to enable the production Ceph
repositories with subscription-manager before using the packages in an
ASYNC advisory's Puddle. With this change, all the latest "1.2.3"
packages will be combined together with the "1.2-async" packages into a
single Puddle.

This has a side-effect advantage in that we can simplify the "externals"
sections of these config files.