Adds cgroups packages and net_cls kernel module to ubuntu systems
This commit is part of the work for http://tracker.ceph.com/issues/12424
which enables the ceph task in ceph/ceph-qa-suite to create and place
osd/mon/mds in control groups.

For ubuntu, kernels for precise and up have CONFIG_CGROUP_* enabled, so
the only thing that it's required is to install cgroup-lite which is a
service that mounts the cgroups virtual FS. The cgroup-bin package will
also be used from the ceph task to manipulate (create/modify/delete)
control groups.

The net_cls subsystem is included as a module (cls_cgroup) and is not
enabled by cgroup-lite, i.e. it has to be explicitly loaded. so we add
it to /etc/modules