users: fixes bug related to ssh key update retries
When updating ssh keys for users we implemented a do until loop that
will account for transient issues when downloading the key from github.
The problem was that if the url fails then the var we were registering
was never created and the 'until' block errors out trying to check it.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>