downstream-setup: separate cleanup and setup tasks into their own files
This is a refactor that makes this role a bit easier to understand and
use. I've simplified main.yml to only include setup.yml and cleanup.yml.
This separation of cleanup and setup related tasks make it a bit easier
to understand.

Also fixes this bug: http://tracker.ceph.com/issues/12693

Signed-off-by: Andrew Schoen <aschoen@redhat.com>