testnode: ensure fuse group exists
On debian 8, uploading the fuse.conf file fails because the fuse group
is no longer present by default. The following debian bug suggests that
testing the fuse group might not be the correct thing to do, but it's
probably best just to keep this as close to debian 7 as possible right
now.

See: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=793927

Signed-off-by: Andrew Schoen <aschoen@redhat.com>