pcp: update upstream packages
PCP's upstream doesn't version its dependencies in all cases. This leads
to situations where, when setting upstream_repo to True, a system ends
up with some old and some new packages. This results in a potentially
broken installation.

This commit causes all installed pcp packages to be upgraded when
upstream_repo is True.

Signed-off-by: Zack Cerza <zack@redhat.com>