common: Change method of detecting beta distro
It was recently discovered that Release Candidate ISOs of RHEL don't output "Beta" or
"Alpha" in `lsb_release` output.  Therefore, RHEL RC Distros were
registering with CDN and not putting beta_repos in place.

This new method registers every RHEL system with CDN, checks if there are
any available matching repos based on ansible_distribution_version, and
either:
  - Unregisters from CDN and runs beta_distros.yml if there are no CDN repos
  - Continues with setting up desired rhsm_repos

I also added some additional working to task descriptions to hopefully
claer up what's going on in this playbook since there are a lot of when
and if statements.

Signed-off-by: David Galloway <dgallowa@redhat.com>