cobbler: Install grub to /dev/sda instead of whatever default is
Apparently on some machines (namely bruuni in Octo), the default disk
returns /dev/nvme0n1 which is not where the OS is installed.

Since we already define which disk the OS should be installed on earlier
in the preseed, it only makes sense to install grub on the same disk.

Signed-off-by: David Galloway <dgallowa@redhat.com>