cobbler_systems: Allow setting post install kernel options
If defined in group_vars, kernel_options_post will be used.  Otherwise,
it defaults to the kernel_options_post in defaults/main.yml.

This fixes serial output during firstboot in Xenial or later (instead of
the repeated "Ubuntu 16.04 [-1;-1f" garbage.

Running `cobbler system edit` with --kopts-post="" is fine.  It just
leaves that field blank if kernel_options_post is empty.

Signed-off-by: David Galloway <dgallowa@redhat.com>