cobbler: Remove DHCP config for NICs if ifup fails in rc.local
An issue was discovered where rc.local bails if a testnode has multiple
NICs cabled but each NIC doesn't have a DHCP reservation.  For example,
some of the magnas have a second NIC cabled but are cabled to a tagged
port on the switch so they can pass traffic via multiple VLANs.

Fixes: http://tracker.ceph.com/issues/22651

Signed-off-by: David Galloway <dgallowa@redhat.com>