pcp: Disable role for now
With the addition of RHEL to Sepia, teuthology will be running
cephlab.yml on unregistered RHEL testnodes.  Since the PCP playbook gets run
before the testnodes playbook, RHEL systems in Sepia won't be registered
to our Satellite yet and PCP installation fails.

We're not currently using PCP so we can disable the role and save some
time and headache.

Signed-off-by: David Galloway <dgallowa@redhat.com>