testnode: Exclude dm devices from list of physical volumes
This was actually happening because when the playbook first runs, the
setup module is run and sees the device mapper devices.  We zap them
later in the playbook but ansible doesn't know that.  We could just
re-run the setup module but this method will instead guarantee we don't
use dm-* devices.

Fixes: https://tracker.ceph.com/issues/23845

Signed-off-by: David Galloway <dgallowa@redhat.com>