cobbler: Add kickstarts that will install to /dev/sdm
The root drive on mero nodes is /dev/sdm.  Since Cobbler is only used to
provision testnodes prior to a FOG image capture, I didn't think it'd be
worth the time/effort to automate this further.

The kickstart will have to be manually selected if the Cobbler Web UI is
used.  I'll automate setting the kickstart when the octo-fog-images
Jenkins job runs.

Signed-off-by: David Galloway <dgallowa@redhat.com>