testnode: remove multipath
multipath can interfere with krbd tests, particularly with
krbd_udev_enumerate.yaml.  Much like udev, multipath opens block
devices behind the scenes, making the block device busy when it
shouldn't be.

None of the lab tests depend on multipath (it's not installed on
ubuntu), so just remove it rather than messing with /etc/multipath.conf
and restarting multipathd.  If the need for multipath comes up, rbd
devices would need to be blacklisted like so:

  blacklist {
          devnode "^rbd[0-9]+"
  }

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>