cobbler: Don't quit rc.local if anything fails
exiting 1 if anything breaks doesn't do us any favors.  What ends up happening is it seems like the system was reimaged but users can't log in because ansible never gets run.  Only admins can log in as the cm user to investigate and even then, since we're outputting to the SOL console, one has to be watching the SOL console for errors when a system is reprovisioned with Cobbler.

If the system fails to get its networking set up, then the ansible tasks won't be able to run anyway.  If ansible fails, there's still the `|| true` at the end so, again, no purpose in having `set -e` anywhere in this script.

Signed-off-by: David Galloway <dgallowa@redhat.com>