generate-fog-csv: Ignore missing mac hostvar
If this is missing in the CSV, FOG will still let you import the host; the MAC field is just blank.  This is fine because we're just using the `[cobbler_managed]` inventory group to define hosts that FOG should manage.  However, some systems (like the refarch switches) are in the `[refarch]` group but don't have MAC addresses.  On top of that, we're not using FOG to reimage the refarch machines so it doesn't matter if the MAC is missing in FOG.

Signed-off-by: David Galloway <dgallowa@redhat.com>