dhcp-server: Only create host declaration if its IP is in subnet we're writing
In the Octo lab, we have multiple subnets with the same domain and VLAN.  This is problematic because in the `dhcp_subnets` dict, we only anticipate there being one cidr.

This change will let you have multiple "subnets" in `dhcp_subnets` without fiddling with the ansible inventory ipvars or macvars.

Signed-off-by: David Galloway <dgallowa@redhat.com>