common: Set release when registering
Because we're running Satellite in a VM in RHV on Gluster, we get really slow I/O performance.  Sadly *well* below the recommended performance: https://access.redhat.com/solutions/3397771

When attempting to set the release in a task later in this tasklist, Satellite would throw 500 errors because it was busy still trying to let the postgres DB know the system had been registered.

Setting the release during registration will save minutes per registration.

Signed-off-by: David Galloway <dgallowa@redhat.com>