Move container tasks to separate role
I would've liked to keep all this in common but there's a chicken and egg situation.

docker and/or podman get installed during the testnode role.  The testnode role can only be run after the common role.  The testnode role is also where some repos are added.

So we need to install docker/podman and configure it after the testnodes role runs.  Since we also want to be able to configure docker/podman on other systems, I couldn't put these tasks in the testnode role.

Signed-off-by: David Galloway <dgallowa@redhat.com>