testnode: Fix CentOS 8 Stream packages list
I guess we were relying on a bug for the past few months. What would happen is centos_8.yml would get pulled in and the `packages` list got populated. Then later when we detect this is CentOS Stream, centos_8_stream.yml would get pulled in which overrode the original `packages` list.

Fixes: https://tracker.ceph.com/issues/55177

Signed-off-by: David Galloway <dgallowa@redhat.com>