common: Remove, clean up, and reinstall Satellite CA cert
This reverts commit ed59d0bb3f6470ecf62fc6af19663d0f674dd84e.

This was originally added because if one upgrades the Satellite server, the katello-ca-consumer-latest RPM actually gets an incremented version number and conflicts with an already installed katello-ca RPM.  `rpm --reinstall` was not sufficient to "upgrade" the package.

We won't ever be upgrading the Satellite server again and the proper way to "reinstall" the Satellite's CA Cert is to remove it, subscription-manager clean, then reinstall as per https://access.redhat.com/solutions/3336921.

Signed-off-by: David Galloway <dgallowa@redhat.com>