Add checkcerts, to use with cron to warn about expiring certs
This originally lived on gitbuilder-archive, and I've moved it,
revamped it, added some args, added some hosts, modified some emails,
ported to Py3, stopped using external programs.  It's quick to run
in default mode where it just reports to the terminal; it'll also
be quiet and only send email about old certs.

The timezone processing is nonexistent on the reported expiry
date; Python timezone handling is a mess.  That could be improved but
not without a deep dive.

Signed-off-by: Dan Mick <dmick@redhat.com>