cephlab_ansible.sh: use scl rh-python38 on CentOS 7
cephlab_ansible.sh runs at the very end of the installation process
during a cobbler install for fog image capture, on first reboot of the
freshly-cobblered system.

Cobbler runs on a CentOS 7 installation today, but its python is too
old to support modern ansible.  The SCL for python 3.8 is installed
on cobbler.  Add code here to, if installed, enable the SCL (by setting
some paths in the trigger script that is executed on the cobbler server
after the installed host reboots; a curl fetch is placed at the end of
/etc/rc.local, and this script runs to finish up all the configuration
of the host for teuthology use.

Signed-off-by: Dan Mick <dmick@redhat.com>