common/tasks/rhel-entitlements.yml: regenerate machine-id
We can't leave /etc/machine-id blank; it breaks things, one of
which is the kernel install, which runs a postinstall script to update
/boot/loader/entries, which does nothing (silently) if there's
nothing in /etc/machine-id.  Since it can come from the dbus id,
and does by default, and there's a command to generate the dbus
id, generate both, dbus first.  This fixes the kernel postinstall.
I don't know if there should be any correlation between
machine-id and the subscription-manager/katello IDs.

Signed-off-by: Dan Mick <dmick@redhat.com>