Remove unnecessary duplicate repo definitions for centos9
CentOS 8 had different .repo files for each major section (BaseOS,
AppStream, etc.).  CentOS 9 has apparently moved to a single file,
centos.repo.  This change 1) removes the management of separate repo files
for BaseOS and AppStream, since those repos are included in centos.repo,
and 2) stops using the perhaps-questionable single baseurl in favor of
the default metalink/mirrors setup

There are errors occurring for teuthology tests on centos9 that may
be related to this, with the errors of the form "<pkg> from <repo>
does not belong to a distupgrade repository".  As near as I can tell,
a "distupgrade repository" is one used only for upgrade, and I can't
find information on how exactly it's indicated, so I don't know if this
change will resolve the error or not.

Signed-off-by: Dan Mick <dmick@redhat.com>